+++ cmd/alleleSplit/toSam.go
@@ -30,7 +30,7 @@ func SnpSearch(samfile string, genotypeVcf string, fOne string, parentOne string
 	sam.WriteHeaderToFileHandle(childTwo, header)
 
 	var i, parentAllele1, parentAllele2 int
-	var target, query, j int64
+	var target, query, j int
 	var ok bool
 	var code uint64
 
@@ -38,7 +38,7 @@ func SnpSearch(samfile string, genotypeVcf string, fOne string, parentOne string
 	var alleles [][]dna.Base
 	for read, done := sam.NextAlignment(samFile); done != true; read, done = sam.NextAlignment(samFile) {
 		parentAllele1, parentAllele2 = 0, 0
-		target = read.Pos - 1
+		target = int(read.Pos - 1)
 		query = 0
 		alleles = vcf.GetAltBases(strings.Split(fmt.Sprintf("%s,%s", gV.Ref, gV.Alt), ","))
 		for i = 0; i < len(read.Cigar); i++ {
@@ -60,20 +60,20 @@ func SnpSearch(samfile string, genotypeVcf string, fOne string, parentOne string
 				//}
 				query += read.Cigar[i].RunLength
 			case 'D':
-				code = vcf.ChromPosToUInt64(int(sampleHash.Fa[read.RName]), int(target))
+				code = vcf.ChromPosToUInt64(int(sampleHash.Fa[read.RName]), target)
 				gV, ok = snpDb[code]
 				if ok {
-					if dna.CountBase(alleles[gV.Samples[sampleHash.GIndex[parentOne]].AlleleOne], dna.Gap) == int(read.Cigar[i].RunLength) && dna.CountBase(alleles[gV.Samples[sampleHash.GIndex[parentOne]].AlleleTwo], dna.Gap) == int(read.Cigar[i].RunLength) {
+					if dna.CountBase(alleles[gV.Samples[sampleHash.GIndex[parentOne]].AlleleOne], dna.Gap) == read.Cigar[i].RunLength && dna.CountBase(alleles[gV.Samples[sampleHash.GIndex[parentOne]].AlleleTwo], dna.Gap) == read.Cigar[i].RunLength {
 						parentAllele1++
 					}
-					if dna.CountBase(alleles[gV.Samples[sampleHash.GIndex[parentTwo]].AlleleOne], dna.Gap) == int(read.Cigar[i].RunLength) && dna.CountBase(alleles[gV.Samples[sampleHash.GIndex[parentTwo]].AlleleTwo], dna.Gap) == int(read.Cigar[i].RunLength) {
+					if dna.CountBase(alleles[gV.Samples[sampleHash.GIndex[parentTwo]].AlleleOne], dna.Gap) == read.Cigar[i].RunLength && dna.CountBase(alleles[gV.Samples[sampleHash.GIndex[parentTwo]].AlleleTwo], dna.Gap) == read.Cigar[i].RunLength {
 						parentAllele1++
 					}
 				}
 				target += read.Cigar[i].RunLength
 			case 'M':
 				for j = 0; j < read.Cigar[i].RunLength; j++ {
-					code = vcf.ChromPosToUInt64(int(sampleHash.Fa[read.RName]), int(target+j))
+					code = vcf.ChromPosToUInt64(int(sampleHash.Fa[read.RName]), target+j)
 					gV, ok = snpDb[code]
 					if ok {
 						if dna.CompareSeqsIgnoreCase([]dna.Base{read.Seq[query+j]}, alleles[gV.Samples[sampleHash.GIndex[parentOne]].AlleleOne]) == 0 && dna.CompareSeqsIgnoreCase([]dna.Base{read.Seq[query+j]}, alleles[snpDb[code].Samples[sampleHash.GIndex[parentOne]].AlleleTwo]) == 0 {
